TTY: translate EIO into end-of-file
Currently when an interactive shell exits then no regular EOF condition
is detected by NST. It looks like the original ST code mostly exits via
SIGCHLD, since they use an asynchronous signal handler for it. In the
adjusted code we pretty consistently encounter a read exception
cosmos::ApiError with errno == EIO.

It looks like this is by design of the PTY as we currently use it. The
read() call for the PTY never returns 0 for end-of-file. Thus catch the
corresponding exception and translate it into end-of-file until I have
some time investigating this topic in more detail. The PTY has some
bells and whistles that might allow to make this clearer to code.

Instead of simply stopping program execution via `exit()` do a proper
return and finish the main() function. This allows us to use tools like
`valgrind` to check for memory and other errors that only fully show if
a proper process exist with destruction of all objects is performed.
This means that closing the terminal can take a tad longer but it should
be irrelevant. We could also change to code to `exit()` if NDEBUG is
set.