types: fix width/height calculation issue
In commit 385b231c8cfcc3e04eceaaf2c7fc412e7e90e77a a bug was introduced
that causes the final column of the terminal not to be drawn. This shows
e.g. when using tmux and trying the search in buffer feature, a single
colored block in the upper right corner is missing.

Fix this by adding the missing 1 to the width() and height()
calculation. It only affects drawScreen() currently. This inclusiveness
of the end coordinate feels a bit strange but let's stick to this for
the moment.