CSIEscape/Term: remove dependency of CSIEscape to StringEscape
StringEscape is only needed for one thing in CSIEscape, and this things
is also duplicate to code already in Term, also kind of private to Term.
So use common code in Term and remove the passing through of
StringEscape into CSIEscape to make it simpler.

Also perform the resetStringEscape() from within Term, not within
StringEscape, thus allowing to make this member function protected in
Term. The function only resets the Term owned escape state, so it's
better invoked there.