Font: default initialize badslant and badweight and turn them into bools
Turning Font and DrawingContext into class members (and no longer having
this as static declared globals) caused a hard to find bug through usage
of undefined data. Previously the Font objects have been zero
initialized by the compiler, now they contain undefined data.

In sanitizeColor() this can cause a reset of the attributes to
DEFAULT_ATTR, meaning all text will be yellow. This only happened
sporadically and especially on Clang builds using newer commits in the
repository. Which is wondrous, because it is not that unlikely that
these two integer variables are non-zero.

Fix this mess by turning the variables into bools, default initializing
them and implementing a reset() member function that properly resets the
Font state. Even in the original ST code this could have happened upon
font re-loading if a badslant/badweight was set initially but should not
be set any more for a newly loaded font.