x11: split-off and refactor font color logic
This moves more logic into the FontColor type. The handling is still
pretty complex, because this contains potentially dynamically loaded
data based on the active X11 colormap.

It's important not to corrupt the standard colors loaded in DrawCtx, but
use copies. Copies are somewhat expensive if colors have to be
recalculated a lot.

The structure is still not ideal since the colors are stored in DrawCtx
while the logic is now a bit more herded in FontColor but a lot also
still in the X11 type. Further reorganization of this is necessary.