X11: carve out color related logic into new ColorManager type
Further simplify the X11 conglomerate by carving out color specific
logic into a new ColorManager type which is a kind of sibbling to the
FontManager.

Along with this also turn the dynamic std::vector colors array into a
statically allocated std::array of the appropriate size.

Minor color related logic still remains in the X11 type which will
be moved in a follow-up commit.