global: rename X11 -> WindowSystem
The class name X11 is overly generic in the NST context. Also filename
and class name don't match. Historically the name stems from the "x.c"
in the original ST source code, which contained half of the program,
actually.

Rename it to WindowSystem which is hopefully clearer for readability and
in any case more harmonized.