StringEscape: fix window title setting bug (and probably many others)
The refactoring in commit 731b8b44 introduced a bug which causes
arguments after the first argument to start with a NULL byte, because
the end iterator is not correctly incremented.

This mostly shows by the terminal title always sticking to the default.
Any other string escape sequences that use more than one parameter are
also affected though.