WindowSystem: fix regression causing missing draw features in column 0
During the refactoring in commit b668c40f some logic got lost that
causes special features of the first character in a line to be ignored.
This shows e.g. when trying to select a range of characters that
includes the first column. The first column text is not reversed as it
should be.

The situation that occurs here is that `glyph.mode.flip()` happens but
the base is not updated it is the first iteration, because the
`base.featuresDiffer()` if block will not trigger to update `base`.