Term: fix regression in tab setup causing screen layout corruption
In commit d1ca9cdf the resize() function was refactored and some
redundant tab setup code was merged. This introduced a bug in
setupTabs() where the new terminal size was not yet visible, so tab
stops for larger column values are not set when the window size is
increased.

This causes screen layout corruption which I experienced in the Mutt
email client when displaying mail threads.

To fix this, update the m_size member earlier. Also drop the unnecessary
`m_tabs.resize()` which happens anyways in `setupTabs()` ->
`clearAllTabs()`.