Term: fix static_assert for Glyph being POD
I'm not completely sure what type the current assert is checking but
it's not the proper nst::Glyph type. As this type isn't POD. It contains
a cosmos::BitMask which has a constructor.

What we actually want to know here is whether Glyph is trivially
copyable, and that is the case. So fix both the type specification and
the type of assertion.