mainLoop(): fix race condition against the child process querying the terminal size
We are currently creating the TTY and its child process before we are
waiting for X11 to setup our window and update our TTY dimensions. This
means that the TTY has an initial (kernel side) size of 0/0.

If the child process is scheduled before we can update the TTY size,
then it can query the terminal size and will see this 0/0 terminal size.
This can cause strange behaviour e.g. in the `less` program which seems
to beleive that there's nothing to output on a zero dimension terminal.

Fix this by waiting for the window to be mapped so we can update the TTY
to the actual initial size. Only then fire up the child process.
Performance wise it could be beneficial to let the child process run
before we even setup the window, but then we would potentially need to
set an initial terminal size (like from nst_config.hxx) that doesn't
match reality (e.g. because the window manager resizes the window right
away). Redimensioning data structures during startup and rearranging
terminal contents due to this could again be negative for performance.

Thus favor correctness over speed for now and do a slow and orderly
startup.