EscapeHandler: fix infinite recursion on CSI repeat sequence
During refactoring this reset statement was optimized out. It turns out
this can lead to an infinite recursion if the last input character is
repeated, because the state handling believes it is still processing a
CSI sequence.

This extra reset is somewhat ugly but there's no smarter quick and
simple solution available it seems.