Term: fix regression in resize() causing defunct scroll up of lines
When the terminal rows are decreased during resize and the decrease of
rows is only 1 then no scrolling up of terminal content occurs any more.
The result is that the last line before the cursor position is
discarded, instead of the first terminal line.

This regression was introduced in commit db72e8749637.