feature: add basic scrollback buffer support
This change roughly corresponds to the original ST patch

    st-scrollback-ringbuffer-0.8.5.diff

It changes the Screen class to replace the first vector dimension (the
rows) into a ring buffer that is used for efficient scrolling. To avoid
having to change the interface a lot this introduces a custom iterator
type that behaves just like before: it iterates over the currently
visible lines on the screen.

The Screen class furthermore gets a couple of new data members to keep
track of the ring buffer and scrollback state. In Term the existing
scrollUp() and scrollDown() need to use some ring buffer specific
shiftViewUp() and shiftViewDown() functions from Screen.

The new config::HISTORY_LEN compile time constant is used to tune
maximum amount of scrollback lines. Additional memory beyond the current
screen content is only allocated if actual scrolling happens.

Scrolling is done using SHIFT + PageUp / Page Down. Upon TTY input
scrolling is reset and the current screen content is displayed again.
The scrollback buffer is only used on the main screen, the alt screen
does not scroll (since it's mainly used for interactive programs that
deal with this themselves, if necessary).