Screen: support preservation of columns upon shrink and unshrink of lines
The full ST reflow patch which supports fully fledged reflow of lines
when changing the terminal window size is horribly complex. I would like
to avoid this complexity.

The part that I really dislike about shrinking a terminal window (which
can quickly happen as a side effect on a tiling window manager) is that
information is possibly lost forever.

This change here attempts to provide an easy transparent solution for
being able to restore line content by simply making the window bigger
again. This works by adding a custom begin()/end()/resize() interface to
LineVector that only returns the columns that are currently in the
screen view, but never shrinks the actual number of columns and thus
preserves content.

Once the screen size is increased again, the preseved columns can
reappear. This still might have a caveeat, if existing lines are edited
on the main screen, while hidden characters from old lines remain. This
is still a TODO to implement, if it is really a pain in practice.