resizing while scrolled back: attempt to keep scrolling position
When the terminal is scrolled back and we resize the window then
scrolling is stopped. This can be counter intuitive when the window size
is e.g. increased because more content of the scroll history is desired.

This change here only works when stopping terminal output via ctrl+s,
because otherwise the shell will redraw the prompt which inevitably
leads to stopScrolling().