feature: support piping the terminal buffer contents to an external program
This introduces a new shortcut Ctrl+Shift+b to open the current terminal
buffer contents (including scroll history) in an external application.
Currently this is set to run `gvim` which is fed the data from stdin via
a pipe.

This is a cheap alternative to implementing a search feature in nst
itself and is more generic. Vim knows very good how to search in a
buffer using regular expressions etc. so let it do the job.