Selection: bugfix: don't wrongly extend word over dummy characters
This assignment is not only unnecessary most of the time, but it is
actually a bug if a dummy character is encountered. Example:

    あ    some word
      ^^^^

When triggering word extension on the sequence of space word delimiters,
then not only the spaces will be selected, but also the wide unicode
character preceding them.

By removing this line this bug will no longer occur. The code is simpler
and we solved the TODO.