Screen: fix handling of empty lines in Screen::asText()
Currently all empty lines are skipped from the asText() output which
yields wrong results, skipping empty lines in the scrollback history.

To fix this properly we need to reverse translate buffer line indices
into screen line indices. While we're at this use a clearer name for
translateLinePos(). Also pass on the information to Screen, if it is the
alternative screen. This helps avoiding the heuristics in
setDimension().

For this fix we now also need to pass the current cursor position to
Screen::asText().