Selection: update selection range upon input when slightly scrolled back
This fixes the following issue:

The screen is scrolled back only a few lines and a selection is
existing. When pressing a key now to return to the current screen then
the selection is not updated to the new coordinates. This shortcoming is
already existing in ST's original ringbuffer scrollback patch. Fix this
by scrolling the selection accordingly when the screen stops scrolling.