Selection: fix jumping selection with KEEP_SCROLL_POSITION
The screen is scrolled back, KEEP_SCROLL_POSITION is configured, a
selection is existing and new input is coming in. When the new input
contains newlines then the terminal logic shifts the current Selection
coordinates as if the selection was on the current screen. Only at the
end the original scroll position is restored. The selection is now
shifted. This is only a visual defect but it is pretty confusing.

To fix this save also the scroll range in the Selection class and
restore it, if necessary. This is not fully ideal, since we are
performing certain changes during input that are unnecessary and need to
be rolled back. It's hard to fix this otherwise though, I believe.