CSIEscape: support title stack operations CSI 22/23
This is an adaption of upstream ST patch st-csi_22_23-0.8.5.diff.

Maintain a stack of window titles that can be pushed to / popped from
via special CSI escape sequences. This is used by some programs to
temporarily change the window title and restore it to its original value
again.

The CSI sequence supports also managing the icon title this way.
Currently we only operate on the window title though, for simplicity.

The terminfo entry is extended to automatically push the current title
when switching to alt screen and popping it when returning to main
screen.

The interactive test is extended to support long command entry and
supports changing the window title and testing push/pop operations.