WindowSystem: fix resetFont() update logic
In commit 3e850281 a regression has been introduced. When resetting to
the default font then various rendering info is not updated properly.
This results in a strange font display especially when returning to the
default font from a much smaller or bigger font.

Fix this by moving the font update logic into its own function that is
called from both zoomFont() and resetFont().