Term: correctly scroll displayed selection when decreasing window height
This is an amendment to the fix in commit fbee7b29. When a selection
exists and the window height is reduced i.e. the screen content needs
to be shifted up, then the selection stays where it is, showing wrong
content.

Fix this by correctly scrolling the selection accordingly as well.