WindowSystem: support shortcut to enter / leave fullscreen mode
This is an adaption of original ST patch st-fullscreen-0.8.5.diff. It
allows to define a shortcut for putting the terminal window into
fullscreen mode and back again.

This is mostly useful for non-tiling window managers that offer no
builtin possibility to display the window in fullscreen without window
borders etc.