WindowSystem: support hiding the mouse cursor on keyboard input
This is an adaption of ST's st-hidecursor patch. If enabled then the
mouse cursor will be hidden from the window once keyboard input occurs,
and will be made visible again one the mouse is moved.

To better differentiate between the text input cursor and the mouse
cursor name the mouse cursor pointer.