StringEscape: properly handle semicolons in title set operations
Currently when a semicolon is part of a title set operation then the
title string stops there, because the rest of the string is handled like
additional, unused parameters.

Add some code for these cases that reconstruct the original title if
necessary.

This can be tested with a sequence like this:

    echo -e "\033]2;title;with;semicolons\a"