Selection: don't scroll selection from the other screen
This fixes a bug that already exists in original ST. If a selection
exists on the other screen then Selection::scroll() operations on the
current screen influence it, which doesn't make sense.

A reproducer for this is as follows:

- select something on the main screen
- enter vim
- press ctrl-z to put VIM in the background
- the selection will have scrolled up by one line

To fix this simply ignore scroll requests if the selection was started
on the other screen.