Term: fix insert mode when entering a wide character at the end of a row
The TODO concern is correct, but only if multi-column wide characters
are involved. When entering such a wide character in insert mode in the
final column, then the current line won't be shifted (since there's not
enough space left). The code after that will move to a new line, but on
this new line no insert mode shifting is done, thus overwriting existing
characters.

To fix this move to newline, if necessary, before insert mode shifting
is considered. Also clear remaining glyphs if the input character does
not fit in the current line. This affects both insert and non-insert
mode.