Term: better compatibility to Xterm regarding alt screen escape codes
This is far more complex than it looks at first. Actually original Xterm
behaves very peculiar regarding the three different escape codes. Even
the documentation PDF does not fully cover the behaviour.

This change should cause the same behaviour as in xterm. Especially the
screen clearing of the alt screen is now handled differently. For code
47 the alt screen is never cleared, for code 1047 the screen is only
cleared when switching away from alt screen and for code 1049 the screen
is only cleared when switching to it.