Term: remove not applicable TODO about autowrap mode disabling
A test against original XTerm shows that the behaviour of (n)st is
equivalent. In interactive.py the behaviour is actually as expected:
with autowrap, when entering text at the last column, the line is
wrapped to the next row. Without autowrap the last column character is
replaced by the new one without the cursor changing position.

When disabling autowrap from within a shell then the behaviour is
strange, though. In bash the characters appear in the first column of
the same row again. In zsh a line wrap occurs but the original line
disappears. This likely has to do with libreadline & friend that don't
support running without autowrap mode.