Term: deleteColsAfterCursor(): fix wide character deletion
This was found during fuzzing. If a wide character is not properly
deleted, then in other spots of the code an attempt can be made to make
an out of bounds access, when encountering the dummy character in column
zero.

Fix this by deleting an extra column when a wide character portion
remains. It's probably not clearly specified what has to happen in such
cases. The alternative would be to replace the dummy character by an
empty field, but this could break the logic just as well.