Selection: use Mode enum instead of Snap plus Flags
The Snap setting is kind of an extension of the actual selection type.
Previously we dropped the selection type and created Flag::RECTANGULAR &
friends for it. This is confusing, since only one selection mode can
ever be active, and the Snap setting on top also influences it.

Thus (re-)introduce a dedicated Selection::Mode that also covers the
Snap settings. This way some things should become clearer.