Selection: harmonize and simplify interface and logic
This is finally an attempt to make the logic of the Selection class
finally clearer and more accessible to the casual reader.

This also introduces some minor behaviour changes e.g. trying SEP_SNAP
or a non-separator character will no longer fall back to WORD_SNAP mode,
but will select nothing.

This also fixes a bug in continueSepSnap() when extending a selection
in a wrapped line scenario. The `delim_pos` variable becomes negative or
too large in such cases and causes an exception to be thrown, resulting
in nst abort.