nst_config: support config file override of SEL_MASKS
With the addition of new selection modes like LINE_RANGE and the
alternative word snap behaviour, the choice of modifiers keys has become
pretty limited. Things are even worse when the window manager consumes
some of the modifiers keys, making it impossible of awkward to use them
for nst.

To allow easy customization towards end users preferences and
environment, support configuration of the modifier keys via the
configuration file.