Screen::asText(): fix end-of-screen detection when scrolled back
screenPos() takes scrolling into accounts, which means that if we're
scrolled back then not the complete buffer will be returned, but only a
part or even an empty string.

This function needs to be independent of scrolling. screenPos() is not
suitable for this, thus make some calculations of our own to make this
independent of scroll state.