nst: support setting theme from command line
It can make sense to create an instance of NST using a specific theme,
without having to rely on the configuration file. To allow for this add
an additional --theme command line switch which takes priority over the
setting found in the configuration file.

Also move the handling of the --dir command line option into a separate
member function to avoid cluttering the main function to much.