IpcHandler protocol: add status for to response
To support further IPC operations that have the possibility to fail, add
an ExitStatus as initial response by the terminal. This breaks
compatibility of the protocol to older versions of NST, but that should
not be an issue, since this is only localhost communication.

Since the seqpacket socket supports message boundaries, implement a
deque queue for sending out messages. We still need to support chunking
for large blocks of data like the screen buffer data.