nst-msg: always use main screen buffer for -d and -D
It turns out that a command line like `nst-msg -d | less` yields varying
results depending on whether nst-msg is quicker in sending out the
request, or less is quicker in switching the terminal to the ALT screen.
This is unfortunate, thus always operate on the main screen to avoid
this race condition.