feature: support changing the theme during runtime via nst-msg
It can make sense to change the color theme during runtime to make
certain contexts bette visible. E.g. a terminal running a remote SSH
session vs. local terminals, or a privilege shell in one terminal and a
regular shell in another.

Changing color is also possible via escape sequences, but this approach
not always be so simple or dynamic as changing it on the fly by sending
an IPC message to NST.

Since this is an IPC message that can fail, the IPC protocol result
evaluation is important. Also two separate messages need to be sent to
NST, the request type and the argument (the theme name).

Since this is a request that changes the drawing state of the terminal,
a new information channel towards the mainLoop is needed, to make it
aware of the need to redraw.

A couple of the main classes need new themeChanged() methods for being
able to react to a change in theme colors. Changing the theme during
runtime is not perfect, because the default colors cannot always be
discerned from the actual colors themselves. This is a drawback that
could only be fixed by adding another bit of information to the Glyph
type, which I don't want to do at the moment.